+++ include/ethercat/Util.hpp
@@ -3,6 +3,10 @@
 #include <array>
 #include <cstdint>
 #include <string>
+#include <boost/describe.hpp>
+#include <boost/mp11.hpp>
+#include <iostream>
+#include <ecrt.h>
 #include <gravity/Transforms.hpp>
 
 namespace gravity
@@ -53,7 +57,7 @@ namespace gravity
     }
 
     template <typename T>
-    std::string to_binary_string(T value)
+    std::string binary_str(T value)
     {
         std::string bits;
         for (int i = sizeof(T) * 8 - 1; i >= 0; --i)
@@ -65,19 +69,40 @@ namespace gravity
         return bits;
     }
 
-    inline std::string hex_str(const uint32_t val, int width)
+    template <typename T>
+    inline std::string hex_str(const T val, int width = sizeof(T) * 2)
     {
+        static_assert(std::is_integral<T>::value, "hex_str requires an integral type");
         std::ostringstream oss;
         oss << "0x"
             << std::uppercase << std::hex
             << std::setw(width) << std::setfill('0')
-            << val;
+            << static_cast<std::make_unsigned_t<T>>(val);
         return oss.str();
     }
 
     inline std::string al_state_str(const uint8_t &al_state)
     {
         return util::tf::enum_str<ec_al_state_t>(static_cast<ec_al_state_t>(al_state));
+    }
+
+    template <typename T>
+    void print_boost_structs(const T &obj)
+    {
+        using namespace boost::describe;
+        using Bases = describe_bases<T, mod_public | mod_protected>;
+        using Members = describe_members<T, mod_public | mod_protected>;
+
+        // walk base classes first
+        boost::mp11::mp_for_each<Bases>([&](auto B)
+                                        {
+        using BaseType = typename decltype(B)::type;
+        print_boost_structs(static_cast<const BaseType&>(obj)); });
+
+        boost::mp11::mp_for_each<Members>(
+            [&](auto D)
+            { std::cout << std::left << std::setw(20) << D.name
+                        << " = " << std::boolalpha << (obj.*D.pointer) << "\n"; });
     }
 
 } // namespace gravity