+++ src/controller/Data.cpp
@@ -1,16 +1,66 @@
 #include "controller/Controller.hpp"
 #include "ethercat/EthStatus.hpp"
 
+bool gravity::Controller::is_op_mode()
+{
+    return domain_state.wc_state == EC_WC_COMPLETE;
+}
+
+bool gravity::Controller::is_running()
+{
+    for (int i = 0; i < motors.size(); i++)
+    {
+        auto status = motor_status[i];
+        status_word_entity sw = decode_status_word(status);
+        if (!sw.servo_running)
+        {
+            // _log->warn("Motor [{}] not running", i);
+            return false;
+        }
+    }
+    return true;
+}
+
+bool gravity::Controller::is_faulted()
+{
+    for (int i = 0; i < motors.size(); i++)
+    {
+        auto status = motor_status[i];
+        status_word_entity sw = decode_status_word(status);
+        if (sw.fault)
+        {
+            _log->warn("Motor [{}] faulted", i);
+            return true;
+        }
+    }
+    return false;
+}
+bool gravity::Controller::is_stopped()
+{
+    // for (int i = 0; i < motors.size(); i++)
+    // {
+    //     auto status = motor_status[i];
+    //     status_word_entity sw = decode_status_word(status);
+    //     if (sw.quick_stop == 0)
+    //     {
+    //         _log->warn("Motor [{}] stopped", i);
+    //         return true;
+    //     }
+    // }
+    return false;
+}
+
 bool gravity::Controller::fetch_current_state(msg::MachineStateInfo &machine_info)
 {
     auto ms = std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch()).count();
     machine_info.ms(ms);
 
     msg::JointStateInfo joint_info;
+    auto master_state = get_master_state();
+    machine_info.al_state(master_state.al_states);
 
     if (master->is_activated())
     {
-        machine_info.al_state(ec_al_state_t::EC_AL_STATE_OP);
         for (int i = 0; i < motors.size(); i++)
         {
             machine_info.motor_state()[i].control_word(motors[i]->control_word->read_pdo());
@@ -34,7 +84,6 @@ bool gravity::Controller::fetch_current_state(msg::MachineStateInfo &machine_inf
     }
     else if (master->is_requested())
     {
-        machine_info.al_state(ec_al_state_t::EC_AL_STATE_PREOP);
         for (int i = 0; i < motors.size(); i++)
         {
             machine_info.motor_state()[i].control_word(motors[i]->control_word->read_sdo());
@@ -55,20 +104,6 @@ bool gravity::Controller::fetch_current_state(msg::MachineStateInfo &machine_inf
             machine_info.joint_state().torque()[i] = config::joint_torque_nm(trq_raw, active_joints[i]);
         }
         return true;
-    }
-    return false;
-}
-
-bool gravity::Controller::is_faulted()
-{
-    for (int i = 0; i < motors.size(); i++)
-    {
-        auto status = motor_status[i];
-        status_word_entity sw = decode_status_word(status);
-        if (sw.fault)
-        {
-            return true;
-        }
     }
     return false;
 }